Updated most of the scripts
Updated:

Eat:
- Ignore when player collides into the sea bed or top sea GameObjects
- When a player eats an enemy the enemies size is added to a score in
the Player script and to a eatenCounter in the GenerateEnemies script
- When a player eats an enemy the camera view and player's acceleration
is scaled appropriately

Enemy:
- Moved the eating functions to a new script called EnemyEat since the
script had to be associated with the enemies head

GenerateEnemies:
- A float called eatenCounter was added to keep track of when all
remaining enemies sizes should be reduced
- A method called ReduceSize was added that goes through the List of
GameObjects of enemies that have been spawned and reduces the sizes of
all that remain in the list if their size is greater than 1
- A method called CheckEatenAllSmall was added that goes through the
List of GameObjects of enemies that have been spawned and sees if and
size 1 or 2 fish remain. Returns false if there are size 1 or 2 fish
else returns true.

Player:
- Added vector based movement with acceleration and drag
- The player's rotation is based on which direction they want to move in
(faces right when moving right and faces left when moving left)

ScreenBound:
- The player's horizontal velocity is set to zero when he hits the
boundary of the background
- The player's vertical velocity is set to zero when he collides with
the sea-bed or top-sea GameObjects

New:

EnemyEat:
- Moved the eat functions for enemies to here as the script needed to be
applied to only the head of the enemy